+++ src/error.rs
@@ -27,7 +27,7 @@ pub enum Error {
     #[error("code id {0}: no such code")]
     UnregisteredCodeId(u64),
 
-    #[error("Unregistered contract address, not present loccaly or on-chain")]
+    #[error("Unregistered contract address, not present locally or on-chain")]
     UnregisteredContractAddress(String),
 }
