+++ src/queries/mod.rs
@@ -1,0 +1,2 @@
+pub mod bank;
+pub mod wasm;
